Refactor different view levels
