+++ core/templates/base_generic.html
@@ -65,7 +65,7 @@
                         </a>
                     </li>
                     <li class="govuk-header__navigation-item {% if request.path == '/forecast/pivotmulti/' %}govuk-header__navigation-item--active{% endif %}">
-                        <a class="govuk-header__link" href="/forecast/pivotmulti/">
+                        <a class="govuk-header__link" href="/forecast/dit/">
                             Forecast
                         </a>
                     </li>
@@ -74,9 +74,10 @@
                             Gifts & Hospitality Register
                         </a>
                     </li>
+                    <!-- TODO Put permissions for CC access conditional here -->
                     <li class="govuk-header__navigation-item {% if request.path == '/forecast/pivot/' %}govuk-header__navigation-item--active{% endif %}">
                         <a class="govuk-header__link" href="/forecast/choose-cost-centre/">
-                            Choose cost centre
+                            Edit cost centre
                         </a>
                     </li>
                     <li class="govuk-header__navigation-item {% if request.path == '/forecast/pivot/' %}govuk-header__navigation-item--active{% endif %}">
