+++ costcentre/forms.py
@@ -42,6 +42,32 @@ class AllCostCentresForm(forms.Form):
     )
 
 
+class DirectorateCostCentresForm(forms.Form):
+    def __init__(self, *args, **kwargs):
+        directorate_code = kwargs.pop('directorate_code')
+        super(DirectorateCostCentresForm, self).__init__(
+            *args,
+            **kwargs,
+        )
+
+        self.base_fields['cost_centre'].queryset = CostCentre.objects.filter(
+            directorate__directorate_code=directorate_code,
+            active=True,
+        )
+
+    cost_centre = forms.ModelChoiceField(
+        queryset=None,
+        widget=Select(),
+    )
+
+    cost_centre.widget.attrs.update(
+        {
+            "class": "govuk-select",
+            "aria-describedby": "cost_centre-hint cost_centre-error",
+        }
+    )
+
+
 class MyCostCentresForm(forms.Form):
     def __init__(self, *args, **kwargs):
         user = kwargs.pop('user')
