+++ forecast/templates/forecast/edit/add.html
