+++ forecast/templates/forecast/edit/choose_cost_centre.html
@@ -1,0 +1,13 @@
+{% extends "wide_content.html" %}
+
+{% block content %}
+    <h3 class="govuk-heading-m">My cost centres</h3>
+    <form method="post" action ="">
+        {% csrf_token %}
+        {% include "core/govt_uk_form.html" %}
+        <input type="submit" value="Choose cost centre" class="govuk-button" data-module="govuk-button"/>
+    </form>
+{% endblock %}
+
+
+
