+++ forecast/templates/forecast/edit/edit_prototype.html
