+++ forecast/templates/forecast/view/dit.html
@@ -1,0 +1,28 @@
+{% extends "wide_content.html" %}
+{% load render_table from django_tables2 %}
+
+{% block content %}
+
+    <div class="govuk-grid-row">
+        <div class="govuk-grid-column-two-thirds">
+            <h1 class="govuk-heading-xl">DIT</h1>
+        </div>
+
+        <div class="govuk-grid-column-one-third directorate-cc-form">
+            <ul class="govuk-list">
+                {% for group in view.groups %}
+                <li>
+                    <a class="govuk-link" href="{% url 'forecast_group' group.group_code %}">{{ group.group_name }}</a>
+                </li>
+                {% endfor %}
+            </ul>
+        </div>
+    </div>
+
+    {% for table in tables %}
+        <div class="tables-container">
+            {% render_table table %}
+        </div>
+    {% endfor %}
+
+{% endblock %}
