+++ forecast/templates/forecast/view/group.html
@@ -1,0 +1,33 @@
+{% extends "wide_content.html" %}
+{% load render_table from django_tables2 %}
+{% load django_bootstrap_breadcrumbs %}
+
+{% block breadcrumbs %}
+    {{ block.super }}
+    {% breadcrumb "DIT forecast" "forecast_dit" %}
+    {% breadcrumb view.group.group_name "" %}
+{% endblock %}
+
+{% block content %}
+    <div class="govuk-grid-row">
+        <div class="govuk-grid-column-two-thirds">
+            <h1 class="govuk-heading-xl">{{ view.group.group_name }}</h1>
+        </div>
+
+        <div class="govuk-grid-column-one-third directorate-cc-form">
+            <ul class="govuk-list">
+                {% for directorate in view.directorates %}
+                <li>
+                    <a class="govuk-link" href="{% url 'forecast_directorate' directorate.directorate_code %}">{{ directorate }}</a>
+                </li>
+                {% endfor %}
+            </ul>
+        </div>
+    </div>
+
+    {% for table in tables %}
+        <div class="tables-container">
+            {% render_table table %}
+        </div>
+    {% endfor %}
+{% endblock %}
